+++ src/Nodes/External/MidiOutput.h
@@ -16,16 +16,16 @@ public:
     std::function<void()> getProcessFunction() const;
     void setProcessFunction(std::function<void()> processFunction);
 
-    std::function<void()> processNote(int channel);
-    std::function<void()> processNoteOn(int channel, std::unordered_set<int> noteSet);
-    std::function<void()> processNoteOff(int channel, std::unordered_set<int> noteSet);
-    std::function<void()> processNotePressure(int channel, std::unordered_set<int> noteSet);
-    std::function<void()> processControl(int channel);
-    std::function<void()> processControlValue(int channel, std::unordered_set<int> controlSet);
-    std::function<void()> processProgram(int channel);
-    std::function<void()> processChannelPressure(int channel);
-    std::function<void()> processPitchBend(int channel);
-    std::function<void()> processAllNotesOff(int channel);
+    std::function<void()> processNote(size_t channel);
+    std::function<void()> processNoteOn(size_t channel, std::unordered_set<uint8> noteSet);
+    std::function<void()> processNoteOff(size_t channel, std::unordered_set<uint8> noteSet);
+    std::function<void()> processNotePressure(size_t channel, std::unordered_set<uint8> noteSet);
+    std::function<void()> processControl(size_t channel);
+    std::function<void()> processControlValue(size_t channel, std::unordered_set<uint8> controlSet);
+    std::function<void()> processProgram(size_t channel);
+    std::function<void()> processChannelPressure(size_t channel);
+    std::function<void()> processPitchBend(size_t channel);
+    std::function<void()> processAllNotesOff(size_t channel);
     std::function<void()> processSongPositionInQuarterNotes();
     std::function<void()> processClock();
     std::function<void()> processStart();
@@ -33,11 +33,11 @@ public:
     std::function<void()> processStop();
 
 protected:
-    void processImpulse(std::unordered_set<int> itemSet, std::function<MidiMessage(int, Sample)> publish);
-    void processImpulse(std::unordered_set<int> itemSet, std::function<MidiMessage(int)> publish);
+    void processImpulse(std::unordered_set<uint8> itemSet, std::function<MidiMessage(uint8, Sample)> publish);
+    void processImpulse(std::unordered_set<uint8> itemSet, std::function<MidiMessage(uint8)> publish);
     void processImpulse(std::function<MidiMessage(Sample)> publish);
     void processImpulse(std::function<MidiMessage()> publish);
-    void processContinuous(std::unordered_set<int> itemSet, std::function<MidiMessage(int, Sample)> publish);
+    void processContinuous(std::unordered_set<uint8> itemSet, std::function<MidiMessage(uint8, Sample)> publish);
     void processContinuous(std::function<MidiMessage(Sample)> publish);
     void processNoLock();
