+++ src/Nodes/Filters/Biquad.h
@@ -20,7 +20,7 @@ public:
     std::shared_ptr<Input> getGain() const;
 
 protected:
-    void setNumOutputChannelsNoLock(int numChannels) override;
+    void setNumOutputChannelsNoLock(size_t numChannels) override;
     void processNoLock() override;
 
 private:
